requirements.txt
numpy==1.17.1
Pillow==6.1.0
